Update kube-batch rep
