+++ Gopkg.lock
@@ -204,7 +204,7 @@
 
 [[projects]]
   branch = "volcano-master"
-  digest = "1:52294fbd6648d468a03845d1d908a8d6e5a86d5cd9fe7f32236d8275721a92df"
+  digest = "1:373e831b25d35980629d8042cf8941f1455cee70e846e2c91064ecb074ce3826"
   name = "github.com/kubernetes-sigs/kube-batch"
   packages = [
     "cmd/kube-batch/app",
@@ -223,6 +223,7 @@
     "pkg/scheduler/actions",
     "pkg/scheduler/actions/allocate",
     "pkg/scheduler/actions/backfill",
+    "pkg/scheduler/actions/enqueue",
     "pkg/scheduler/actions/preempt",
     "pkg/scheduler/actions/reclaim",
     "pkg/scheduler/api",
@@ -244,7 +245,7 @@
     "pkg/version",
   ]
   pruneopts = "UT"
-  revision = "5e9977d68a13938fbb6d7dda564da0dca1b1c98c"
+  revision = "b880cece03e0c1b9c4efb20ca82e3b091a20c43b"
   source = "https://github.com/volcano-sh/kube-batch"
 
 [[projects]]
@@ -1076,17 +1077,21 @@
     "k8s.io/apimachinery/pkg/util/strategicpatch",
     "k8s.io/apimachinery/pkg/util/uuid",
     "k8s.io/apimachinery/pkg/util/validation",
+    "k8s.io/apimachinery/pkg/util/validation/field",
     "k8s.io/apimachinery/pkg/util/wait",
     "k8s.io/apimachinery/pkg/watch",
     "k8s.io/apiserver/pkg/util/flag",
     "k8s.io/client-go/discovery",
     "k8s.io/client-go/discovery/fake",
     "k8s.io/client-go/informers",
+    "k8s.io/client-go/informers/core/v1",
+    "k8s.io/client-go/informers/scheduling/v1beta1",
     "k8s.io/client-go/kubernetes",
     "k8s.io/client-go/kubernetes/scheme",
     "k8s.io/client-go/kubernetes/typed/admissionregistration/v1beta1",
     "k8s.io/client-go/kubernetes/typed/core/v1",
     "k8s.io/client-go/listers/core/v1",
+    "k8s.io/client-go/listers/scheduling/v1beta1",
     "k8s.io/client-go/plugin/pkg/client/auth/gcp",
     "k8s.io/client-go/rest",
     "k8s.io/client-go/testing",
