+++ installer/chart/volcano/config/kube-batch.conf
@@ -1,4 +1,4 @@
-actions: "reclaim, allocate, backfill, preempt"
+actions: "enqueue, reclaim, allocate, backfill, preempt"
 tiers:
 - plugins:
   - name: priority
